+++ demo/src/App.js
@@ -3,7 +3,7 @@ import { Routes, Route } from "react-router-dom";
 import DisplayPlants from "./pages/DisplayPlants";
 import Login from "./pages/Login";
 import Account from "./pages/Account";
-import Register from "./pages/Register"; // Import Register page
+import Register from "./pages/Register";
 
 function App() {
   return (
@@ -13,7 +13,7 @@ function App() {
         <Route path="/display-plants" element={<DisplayPlants />} />
         <Route path="/login" element={<Login />} />
         <Route path="/account" element={<Account />} />
-        <Route path="/register" element={<Register />} /> {/* Add Register route */}
+        <Route path="/register" element={<Register />} />
       </Routes>
     </div>
   );
